Updated
This reads a file now using the inputs for the main method. Gonna keep refining it and it needs to be integrated with the update code and the proper place for where the files are stored. (Right now its just under the c:/ drive and that only works for windows.)